Move background services tests to new class
